+++ Sources/SwiftExtras/SwiftUI/MultiSelectView.swift
@@ -102,7 +102,7 @@ public struct MultiSelectView<Label: View>: View {
 @available(iOS 17, macOS 14, tvOS 17, visionOS 1, watchOS 10, *)
 #Preview {
     @Previewable @State var selectedItems: [String] = []
-    var items = ["star", "person", "rainbow"]
+    let items = ["star", "person", "rainbow"]
 
     MultiSelectView(
         sourceItems: items,